Inline NvChad into config
I've been running NvChad for almost a year now and I really enjoy it.
I'm running a fork at v2.0 and I haven't felt the urge to pull changes
from upstream. I'm including the NvChad fork in my dotfiles as a
submodule and it's been a pain to maintain as I'm not used to working
with submodules. Since I haven't been keeping up with upstream, at this
point I'd like to just inline NvChad and my customizations and treat it
as my neovim config.